fixed to avoid adding already exist tweets

+++ twepy.py
@@ -1,4 +1,5 @@
 import tweepy
+from os import path
 
 with open("secrets.txt", "r") as s:
     secrets = s.readlines()
@@ -13,6 +14,27 @@ auth.set_access_token(access_token, access_token_secret)
 api = tweepy.API(auth)
 
 public_tweets = api.search_30_day("dev30", "@LCWaikiki")
-with open("tweets.txt", "w") as f:
-    for tweet in public_tweets:
-        f.write(tweet.text + "\n============\n")
+
+count = 0
+if path.isfile("tweets.txt"):
+    print("file already exist")
+    with open("tweets.txt", "r") as f:
+        old_data = f.read()
+    old_data = old_data.split("\n============\n")
+    with open("tweets.txt", "a") as f:
+        for tweet in public_tweets:
+            if tweet.text in old_data:
+                print("tweet already exist in the file")
+                print(tweet.text)
+                continue
+            else:
+                count +=1
+                f.write(tweet.text + "\n============\n")
+else:
+    print("file does not exist")
+    with open("tweets.txt", "w") as f:
+        for tweet in public_tweets:
+            count +=1
+            f.write(tweet.text + "\n============\n")
+
+print(count, "number of tweets added to file")
